Nova vista i canvis en el perfil. Encara no funciona busqueda usuari amb JSON
